backend/requirements.txt

fastapi
uvicorn[standard]
pydantic
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
python-dotenv
google-generativeai
pinecone-client
langchain
pytest
